updated README file
