+++ test/test_evaluators.py
@@ -5,8 +5,8 @@ import pathlib
 from PIL import Image
 
 from vision_evaluation.evaluators import AveragePrecisionEvaluator, F1ScoreEvaluator, TopKAccuracyEvaluator, ThresholdAccuracyEvaluator, MeanAveragePrecisionEvaluatorForSingleIOU, EceLossEvaluator, \
-    PrecisionEvaluator, MeanAveragePrecisionAtK, RecallEvaluator, TagWiseAccuracyEvaluator, TagWiseAveragePrecisionEvaluator, \
-    MeanAveragePrecisionNPointsEvaluator, PrecisionRecallCurveNPointsEvaluator, BalancedAccuracyScoreEvaluator, CocoMeanAveragePrecisionEvaluator, BleuScoreEvaluator, METEORScoreEvaluator, \
+    PrecisionEvaluator, RecallEvaluator, TagWiseAccuracyEvaluator, TagWiseAveragePrecisionEvaluator, \
+    MeanAveragePrecisionNPointsEvaluator, BalancedAccuracyScoreEvaluator, CocoMeanAveragePrecisionEvaluator, BleuScoreEvaluator, METEORScoreEvaluator, \
     ROUGELScoreEvaluator, CIDErScoreEvaluator, SPICEScoreEvaluator, RocAucEvaluator, MeanIOUEvaluator, ForegroundIOUEvaluator, BoundaryMeanIOUEvaluator, BoundaryForegroundIOUEvaluator, \
     L1ErrorEvaluator, GroupWiseEvaluator, MeanLpErrorEvaluator
 from vision_evaluation.prediction_filters import TopKPredictionFilter, ThresholdPredictionFilter
@@ -839,122 +839,3 @@ class TestMeanLpErrorEvaluator(unittest.TestCase):
         evaluator_l2.add_predictions(predictions=self.PREDICTIONS[5:], targets=self.TARGETS[5:])
         report = evaluator_l2.get_report()
         self.assertAlmostEqual(report[evaluator_l2._get_id()], np.sqrt(10) / 10, places=4)
-
-
-class TestInformationRetrievalMetrics(unittest.TestCase):
-    PREDICTIONS = [np.array([[5, 4, 3, 2, 1]]),
-                   np.array([[5, 4, 3, 2, 1]]),
-                   np.array([[1, 2, 3, 4, 5]]),
-                   np.array([[5, 4, 3, 2, 1]]),
-                   np.array([[5, 4, 3, 2, 1],
-                             [5, 4, 3, 2, 1]]),
-                   np.array([[5, 4, 3, 2, 1],
-                             [5, 4, 3, 2, 1]]),
-                   np.array([[1]]),
-                   np.array([[2],
-                             [3]])]
-    TARGETS = [np.array([[1, 1, 0, 0, 1]]),
-               np.array([[1, 1, 0, 0, 1]]),
-               np.array([[1, 0, 0, 1, 1]]),
-               np.array([[0, 0, 0, 0, 1]]),
-               np.array([[0, 0, 0, 0, 1],
-                         [0, 0, 0, 0, 1]]),
-               np.array([[1, 0, 0, 0, 1],
-                         [0, 0, 0, 0, 1]]),
-               np.array([[1]]),
-               np.array([[1],
-                         [0]])]
-
-    def test_recall_at_k(self):
-        ks = [6, 8, 6, 6, 6, 6, 4, 4]
-        expectations = [[0, 0.33333, 0.66666, 0.66666, 0.66666, 1.0],
-                        [0, 0.33333, 0.66666, 0.66666, 0.66666, 1.0, 1.0, 1.0],
-                        [0, 0.33333, 0.66666, 0.66666, 0.66666, 1.0],
-                        [0, 0, 0, 0, 0, 1.0],
-                        [0, 0, 0, 0, 0, 1.0],
-                        [0, 0.25, 0.25, 0.25, 0.25, 1.0],
-                        [0, 1.0, 1.0, 1.0],
-                        [0, 0.5, 0.5, 0.5]]
-        assert len(self.PREDICTIONS) == len(self.TARGETS) == len(ks) == len(expectations)
-        for preds, tgts, exps, k in zip(self.PREDICTIONS, self.TARGETS, expectations, ks):
-            for i in range(k):
-                recall_eval = RecallEvaluator(TopKPredictionFilter(i))
-                recall_eval.add_predictions(preds, tgts)
-                self.assertAlmostEqual(recall_eval.get_report(average='samples')[f"recall_top{i}"], exps[i], places=4)
-
-    def test_precision_at_k(self):
-        ks = [6, 8, 6, 6, 6, 6, 4, 4]
-        expectations = [[0, 1.0, 1.0, 0.66666, 0.5, 0.6],
-                        [0, 1.0, 1.0, 0.66666, 0.5, 0.6, 0.6, 0.6],
-                        [0, 1.0, 1.0, 0.66666, 0.5, 0.6],
-                        [0, 0, 0, 0, 0, 0.2],
-                        [0, 0, 0, 0, 0, 0.2],
-                        [0, 0.5, 0.25, 0.16666, 0.125, 0.3],
-                        [0, 1.0, 1.0, 1.0],
-                        [0, 0.5, 0.5, 0.5]]
-        assert len(self.PREDICTIONS) == len(self.TARGETS) == len(ks) == len(expectations)
-        for preds, tgts, exps, k in zip(self.PREDICTIONS, self.TARGETS, expectations, ks):
-            for i in range(k):
-                precision_eval = PrecisionEvaluator(TopKPredictionFilter(i))
-                precision_eval.add_predictions(preds, tgts)
-                self.assertAlmostEqual(precision_eval.get_report(average='samples')[f"precision_top{i}"], exps[i], places=4)
-
-    def test_precision_recall_curve(self):
-        predictions = [np.array([[5, 4, 3, 2, 1]]),
-                       np.array([[1, 3, 2, 5, 4]]),
-                       np.array([[5, 4, 3, 2, 1]]),
-                       np.array([[5, 4, 3, 2, 1],
-                                 [5, 4, 3, 2, 1]])]
-        targets = [np.array([[0, 0, 0, 0, 1]]),
-                   np.array([[1, 0, 0, 0, 0]]),
-                   np.array([[1, 0, 0, 0, 1]]),
-                   np.array([[0, 0, 0, 0, 1],
-                             [1, 0, 0, 0, 1]])]
-
-        expectations = [np.array([0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 1., ]),
-                        np.array([0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 1., ]),
-                        np.array([0.4, 0.4, 0.4, 0.4, 0.4, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, ]),
-                        np.array([0.3, 0.3, 0.3, 0.3, 0.3, 0.6, 0.6, 0.6, 0.6, 0.6, 1., ])]
-        assert len(predictions) == len(targets) == len(expectations)
-        for preds, tgts, exps in zip(predictions, targets, expectations):
-            n_points = 11
-            evaluator = PrecisionRecallCurveNPointsEvaluator(n_points)
-            evaluator.add_predictions(predictions=preds, targets=tgts)
-            self.assertAlmostEqual(np.sum(np.abs(evaluator.get_report(average='samples')[f"PR_Curve_{n_points}_point_interp"] - exps)), 0.0, places=4)
-
-    def test_mean_average_precision_at_k_evaluator(self):
-        targets = [np.array([[1, 0, 1, 1],
-                             [1, 0, 0, 1]]),
-                   np.array([[1, 0, 1, 1]]),
-                   np.array([[1, 0, 0, 1]]),
-                   np.array([[]]),
-                   np.array([[1, 0, 1, 1]]),
-                   np.array([[1, 0, 1, 1]]),
-                   np.array([[1, 0, 1, 1]]),
-                   np.array([[1, 0, 1, 1]]),
-                   np.array([[1, 0, 1, 1]]),
-                   np.array([[1, 0, 1, 1]]),
-                   np.array([[0, 0, 0, 0]]),
-                   np.array([[1, 0, 1]])]
-        predictions = [np.array([[5, 4, 3, 2],
-                                 [5, 4, 3, 2]]),
-                       np.array([[5, 4, 3, 2]]),
-                       np.array([[5, 4, 3, 2]]),
-                       np.array([[]]),
-                       np.array([[2, 3, 5, 4]]),
-                       np.array([[4, 2, 3, 5]]),
-                       np.array([[4, 2, 3, 5]]),
-                       np.array([[2, 3, 5, 4]]),
-                       np.array([[2, 3, 5, 4]]),
-                       np.array([[2, 3, 5, 4]]),
-                       np.array([[2, 3, 5, 4]]),
-                       np.array([[2, 3, 5]])]
-        rank = [4, 4, 4, 4, 4, 4, 3, 3, 5, 2, 4, 4]
-        expectations = [0.77777, 0.80555, 0.75, 0.0, 0.91666, 1.0, 1.0, 0.66666, 0.91666, 1.0, 0.0, 0.83333]
-
-        assert len(targets) == len(predictions) == len(rank) == len(expectations)
-
-        for preds, tgts, exps, k in zip(predictions, targets, expectations, rank):
-            evaluator = MeanAveragePrecisionAtK(k)
-            evaluator.add_predictions(preds, tgts)
-            self.assertAlmostEqual(evaluator.get_report(average='samples')[f"map_top{k}"], exps, places=4)
