+++ vision_evaluation/evaluators.py
@@ -269,68 +269,7 @@ class PrecisionEvaluator(MemorizingEverythingEvaluator):
         return super(PrecisionEvaluator, self).calculate_score(average=average, filter_out_zero_tgt=filter_out_zero_tgt)
 
     def _calculate(self, targets, predictions, average):
-        """
-        There is a special case that needs to be handled appropriately.
-        Due to filtering conditions, there are occasions where only 1 class remains in targets/predictions and sklearn interpretes this as an invalid configuration for multilabel.
-        Since when average == 'samples' is only calculated for multilabel, sm.recall_score throws an exception. e.g.:
-        predictions = np.array([[True]])
-        targets = np.array([[True]])
-        recall_score(targets, predictions, average='samples') throws the following error:
-        ValueError: Samplewise metrics are not available outside of multilabel classification.
-        """
-        if targets.shape[1] == 1 and average == 'samples':
-            targets = np.append(targets, np.zeros((targets.shape[0], 1)), axis=1)
-            predictions = np.append(predictions, np.zeros((predictions.shape[0], 1)), axis=1)
         return sm.precision_score(targets, predictions, average=average)
-
-
-class MeanAveragePrecisionAtK(MemorizingEverythingEvaluator):
-    """
-    MeanAveragePrecision @ K as defined here:
-    https://stackoverflow.com/questions/54966320/mapk-computation
-    https://medium.com/@misty.mok/how-mean-average-precision-at-k-map-k-can-be-more-useful-than-other-evaluation-metrics-6881e0ee21a9
-    MAP@K is the average of AveP(K) over all queries. Hence it uses average="samples".
-    AveP(K) =  Num / Den, where:
-    Num = sum_i^k P(i) * rel(i), where:
-    P(i) is the Precision @ i
-    rel(i) is an indicator function where rel(i) = 1 if position i is relevant, and rel(i) = 0 otherwise.
-    Den = min(K, number of relevant images)
-    """
-
-    def __init__(self, k):
-        super(MeanAveragePrecisionAtK, self).__init__()
-        self.k = k
-
-    def _get_id(self):
-        return f'map_top{self.k}'
-
-    def calculate_score(self, average='samples'):
-        assert average == 'samples'
-        if self.k == 0:
-            return 0.0
-        return self._calculate()
-
-    def _calculate(self):
-        assert self.all_predictions.shape == self.all_targets.shape
-        if self.all_predictions.size == 0:
-            return 0.0
-        return np.mean([self._average_precision_at_k(preds, targets) for preds, targets in zip(self.all_predictions, self.all_targets)])
-
-    def _average_precision_at_k(self, predictions, targets):
-        total_pos_gt = np.sum(targets)
-        if total_pos_gt == 0:
-            return 0.0
-        rank = min(self.k, len(predictions))
-        top_k_pred_indices = np.argpartition(-predictions, rank - 1)[:rank]
-        # sort score, gt by top_k
-        targets = targets[top_k_pred_indices]
-        sum = 0.0
-        num_hits = 0.0
-        for idx, tgt in enumerate(targets):
-            if tgt:
-                num_hits += 1.0
-                sum += num_hits / (idx + 1.0)
-        return sum / min(rank, total_pos_gt)
 
 
 class RecallEvaluator(MemorizingEverythingEvaluator):
@@ -345,18 +284,6 @@ class RecallEvaluator(MemorizingEverythingEvaluator):
         return f'recall_{self.prediction_filter.identifier}'
 
     def _calculate(self, targets, predictions, average):
-        """
-        There is a special case that needs to be handled appropriately.
-        Due to filtering conditions, there are occasions where only 1 class remains in targets/predictions and sklearn interpretes this as an invalid configuration for multilabel.
-        Since when average == 'samples' is only calculated for multilabel, sm.recall_score throws an exception. e.g.:
-        predictions = np.array([[True]])
-        targets = np.array([[True]])
-        recall_score(targets, predictions, average='samples') throws the following error:
-        ValueError: Samplewise metrics are not available outside of multilabel classification.
-        """
-        if targets.shape[1] == 1 and average == 'samples':
-            targets = np.append(targets, np.zeros((targets.shape[0], 1)), axis=1)
-            predictions = np.append(predictions, np.zeros((predictions.shape[0], 1)), axis=1)
         return sm.recall_score(targets, predictions, average=average)
 
 
@@ -871,30 +798,6 @@ class MeanAveragePrecisionNPointsEvaluator(PrecisionRecallCurveMixin, Memorizing
 
     def _get_id(self):
         return f'mAP_{self.n_points}_points'
-
-
-class PrecisionRecallCurveNPointsEvaluator(PrecisionRecallCurveMixin, MemorizingEverythingEvaluator):
-    """
-    N-point interpolatedprecision-recall curve, averaged over samples
-    """
-
-    def calculate_score(self, average='samples', filter_out_zero_tgt=False):
-        assert average == 'samples'
-        return super(PrecisionRecallCurveNPointsEvaluator, self).calculate_score(average=average, filter_out_zero_tgt=filter_out_zero_tgt)
-
-    def _calculate(self, targets, predictions, average):
-        assert average == 'samples'
-        n_samples = predictions.shape[0]
-        recall_thresholds = np.linspace(1, 0, self.n_points, endpoint=True).tolist()
-        precision_averaged = np.zeros(self.n_points)
-        for i in range(n_samples):
-            precision_interp = self._calc_precision_recall_interp(predictions[i, :], targets[i, :], recall_thresholds)
-            precision_averaged += precision_interp
-        precision_averaged /= n_samples
-        return precision_averaged
-
-    def _get_id(self):
-        return f'PR_Curve_{self.n_points}_point_interp'
 
 
 class ImageCaptionEvaluatorBase(Evaluator):
